# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce SaaS API",
        "description": "API REST para la plataforma ecommerce multi-tenant SaaS. Incluye endpoints para catálogo, carrito, checkout, recomendaciones y cotizaciones.",
        "contact": {
            "name": "Soporte API",
            "email": "api@ecommerce.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Listar notificaciones",
                "description": "Retorna las notificaciones del usuario autenticado, paginadas.",
                "operationId": "0f031f66cd8c52bf73e89fb2d050ad83",
                "responses": {
                    "200": {
                        "description": "Notificaciones paginadas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar como leída",
                "description": "Marca una notificación específica como leída.",
                "operationId": "e8ec77e64be528452b6b9ee62d832d70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la notificación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación marcada como leída"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/read-all": {
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar todas como leídas",
                "description": "Marca todas las notificaciones no leídas como leídas.",
                "operationId": "e873443612632a319ae0f88db368dd7a",
                "responses": {
                    "200": {
                        "description": "Todas marcadas como leídas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Conteo de no leídas",
                "description": "Retorna el número de notificaciones no leídas.",
                "operationId": "5120217e8b1a75b2c61e53684a38c045",
                "responses": {
                    "200": {
                        "description": "Conteo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}